feat: Implement Quick Reply Manager for managing quick replies
